Rearranged code and made a common "precompiled" header

+++ Pad02/Text.cpp
@@ -1,7 +1,6 @@
+#include "precompiled.h"
 #include "Text.h"
 
-#include <array>
-#include <dwrite_2.h>
 
 #include "Graphic.h"
 
@@ -9,18 +8,18 @@ Pad02::Text::Text()
 {
 	HRESULT hr = S_OK;
 	hr = DWriteCreateFactory(DWRITE_FACTORY_TYPE_ISOLATED,
-	                         _uuidof(IDWriteFactory),
-	                         reinterpret_cast<IUnknown**>(pWriteFactory.GetAddressOf()));
+		_uuidof(IDWriteFactory),
+		reinterpret_cast<IUnknown**>(pWriteFactory.GetAddressOf()));
 	ValidateResult(hr, "failed to create direct write factory");
 	hr = pWriteFactory->CreateTypography(&pTypography);
 	ValidateResult(hr, "failed to create typography");
-	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_2, 2});
+	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{ DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_2, 2 });
 	ValidateResult(hr, "failed to add font feature SS02");
-	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{DWRITE_FONT_FEATURE_TAG_SLASHED_ZERO, 1});
+	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{ DWRITE_FONT_FEATURE_TAG_SLASHED_ZERO, 1 });
 	ValidateResult(hr, "failed to add font feature SLASHED_ZERO");
-	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_7, 1});
+	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{ DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_7, 1 });
 	ValidateResult(hr, "failed to add font feature ROUND PARENTHESIS");
-	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_8, 1});
+	hr = pTypography->AddFontFeature(DWRITE_FONT_FEATURE{ DWRITE_FONT_FEATURE_TAG_STYLISTIC_SET_8, 1 });
 	ValidateResult(hr, "failed to add font feature ROUND PARENTHESIS");
 
 
@@ -34,10 +33,10 @@ Pad02::Text::Text()
 	hr = pWriteFactory->CreateFontFileReference(L"C:\\Code\\MonoLisa\\ttf\\MonoLisaNormal.ttf", nullptr, &pFontFile);
 	ValidateResult(hr, "failed to create reference to font");
 
-	std::array<IDWriteFontFile*, 1> fontFiles{{pFontFile.Get()}};
+	std::array<IDWriteFontFile*, 1> fontFiles{ {pFontFile.Get()} };
 	hr = pWriteFactory->CreateFontFace(
 		DWRITE_FONT_FACE_TYPE_TRUETYPE,
-		fontFiles.size(), fontFiles.data()
+		gsl::narrow_cast<UINT32>(fontFiles.size()), fontFiles.data()
 		, 0, DWRITE_FONT_SIMULATIONS_NONE, &pFontFace);
 	ValidateResult(hr, "failed to create font face");
 
@@ -54,30 +53,30 @@ auto Pad02::Text::CreateTextFormat(const std::wstring fontName, float fontSize) 
 {
 	HRESULT hr = S_OK;
 	hr = pWriteFactory->CreateTextFormat(fontName.c_str(),
-	                                     nullptr,
-	                                     DWRITE_FONT_WEIGHT_NORMAL,
-	                                     DWRITE_FONT_STYLE_NORMAL,
-	                                     DWRITE_FONT_STRETCH_NORMAL,
-	                                     fontSize,
-	                                     L"",
-	                                     pTextFormat.GetAddressOf());
+		nullptr,
+		DWRITE_FONT_WEIGHT_NORMAL,
+		DWRITE_FONT_STYLE_NORMAL,
+		DWRITE_FONT_STRETCH_NORMAL,
+		fontSize,
+		L"",
+		pTextFormat.GetAddressOf());
 	ValidateResult(hr, "failed to create text format");
 }
 
 auto Pad02::Text::CreateTextLayout(const std::wstring& text, float maxWidth,
-                                   float maxHeight) const -> ComPtr<IDWriteTextLayout>
+	float maxHeight) const -> ComPtr<IDWriteTextLayout>
 {
 	ComPtr<IDWriteTextLayout> pTextLayout = nullptr;
 	pWriteFactory->CreateTextLayout(
-		text.c_str(), text.size(),
+		text.c_str(), gsl::narrow_cast<UINT32>(text.size()),
 		pTextFormat.Get(),
 		maxWidth, maxHeight,
 		pTextLayout.GetAddressOf());
-	
+
 	return pTextLayout;
 }
 
-		
+
 auto Pad02::Text::HitTestPoint(float x, float y) -> DWRITE_HIT_TEST_METRICS
 {
 	return DWRITE_HIT_TEST_METRICS{};
